requirements.txt
langchain
numpy
openai
pandas
prettytable
openapi-schema-pydantic
streamlit
python-dotenv